- Add GPU picking - Add arrow mesh - Make merge use offset - Add rotateVertices

+++ ui/include/rawrbox/ui/elements/input.hpp
@@ -99,7 +99,7 @@ namespace rawrbox {
 		[[nodiscard]] virtual const rawrbox::Color& getBackgroundColor() const;
 
 		virtual void setFont(rawrbox::Font* font);
-		virtual void setFont(const std::filesystem::path& font, int size = 11);
+		virtual void setFont(const std::filesystem::path& font, uint16_t size = 11);
 		[[nodiscard]] virtual rawrbox::Font* getFont() const;
 
 		[[nodiscard]] virtual bool empty() const;
